CNAM-94 filtering main

+++ src/main/scala/fr/polytechnique/cmap/cnam/filtering/utilities/TransformerHelper.scala
@@ -1,11 +1,11 @@
 package fr.polytechnique.cmap.cnam.filtering.utilities
 
-import org.apache.spark.sql.{Column}
-import org.apache.spark.sql.types.{LongType, TimestampType}
+import org.apache.spark.sql.Column
+import org.apache.spark.sql.types.{DoubleType, TimestampType}
 
 object TransformerHelper {
 
   def getMeanTimestampColumn(timestamp1: Column, timestamp2: Column): Column = {
-    ((timestamp1.cast(LongType) + timestamp2.cast(LongType)) / 2).cast(TimestampType)
+    ((timestamp1.cast(DoubleType) + timestamp2.cast(DoubleType)) / 2D).cast(TimestampType)
   }
 }
